+++ src/Machina/Drivers/Driver.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Machina.Users;
+using Machina;
 using Machina.Types.Geometry;
 using Machina.Descriptors.Cursors;
-
+using Machina.Types.Data;
 
 namespace Machina.Drivers
 {